psi: package test fixtures spell their Wrapping-to-u64 returns
2396af9ada rejects a store that moves a typed value to another integer
type or arithmetic policy. `add_u64(left: u64 in Wrapping, right: u64 in
Wrapping) -> u64` returned the Wrapping sum directly, so its package
failed checked compilation with "store in machine `add_u64` state `entry`
terminal expression implicitly converts `u64 in Wrapping` to `u64`". That
body appears in the arithmetic-kernels package fixture and in the four
conflict_probe sources of the package-manager capability-conflict
transaction tests.

Each body now spells `(right + left) as u64` or `(left + right) as u64`.
Dropping the policy is the spelling the rule asks for, and the u64 return
keeps the public signature and its review rows unchanged.

In `cargo nextest run -p package-manager -p topology-plan -p omega
--no-fail-fast`, the 7 package-manager tests that failed on this
diagnostic now pass: compiler_review_evidence's local fixture sweep, the 4
package_reconstruction_question tests over graph-workbench, the recorded
remote-pin replay through the local SSH transport, and the
capability-conflict transaction test. The pinned remote arithmetic-kernels
revisions keep the old body; REMOTE_PINS.md already records that remote
acceptance needs migrated revisions.

+++ omega-rust/omega/packages/manager/tests/capability_conflicts/transaction/conflict_behavior.rs
@@ -167,7 +167,7 @@ pub(super) fn assert_comparison_limits_and_risk_classes(
         r#"boundary data PlatformToken;
 
 pub machine add_u64(left: u64 in Wrapping, right: u64 in Wrapping) -> u64 {
-    left + right
+    (left + right) as u64
 }
 "#,
     );
